Refactor bash aliases and functions
