+++ bash/aliases/docker.aliases
@@ -1,0 +1,8 @@
+# Docker aliases  
+alias dps='docker ps'  
+alias dpsa='docker ps -a'  
+alias di='docker images'  
+alias db='docker build'  
+alias dr='docker run'  
+alias drm='docker rm $(docker ps -a -q)'  
+alias drmi='docker rmi $(docker images -q)'
