+++ bash/aliases/git.aliases
@@ -1,0 +1,11 @@
+# Git aliases  
+alias gs='git status'  
+alias ga='git add'  
+alias gb='git branch'  
+alias gc='git commit'  
+alias gco='git checkout'  
+alias gl='git log --oneline --graph --all --decorate'  
+alias gd='git diff'  
+alias gp='git push'  
+alias gpo='git push origin'  
+alias gpl='git pull'
