+++ bash/aliases/kubernetes.aliases
@@ -1,0 +1,5 @@
+# Kubernetes aliases  
+alias k='kubectl'  
+alias kgp='kubectl get pods'  
+alias kgd='kubectl get deployments'  
+
