+++ bin/checkPrices.py
@@ -1,0 +1,47 @@
+#!/usr/bin/env python3  
+import requests  
+import json  
+from tabulate import tabulate  
+  
+def build_pricing_table(json_data, table_data):  
+    for item in json_data['Items']:  
+        meter = item['meterName']  
+        table_data.append([  
+            item['armSkuName'],  
+            item['retailPrice'],  
+            item['unitOfMeasure'],  
+            item['armRegionName'],  
+            meter,  
+            item['productName']  
+        ])  
+  
+def main():  
+    table_data = []  
+    table_data.append([  
+        'SKU',  
+        'Retail Price',  
+        'Unit of Measure',  
+        'Region',  
+        'Meter',  
+        'Product Name'  
+    ])  
+  
+    api_url = "https://prices.azure.com/api/retail/prices"  
+    query = ("armRegionName eq 'southcentralus' and armSkuName eq 'Standard_NP20s' "  
+             "and priceType eq 'Consumption' and contains(meterName, 'Spot')")  
+    response = requests.get(api_url, params={'$filter': query})  
+    json_data = json.loads(response.text)  
+  
+    build_pricing_table(json_data, table_data)  
+    nextPage = json_data['NextPageLink']  
+  
+    while nextPage:  
+        response = requests.get(nextPage)  
+        json_data = json.loads(response.text)  
+        nextPage = json_data.get('NextPageLink', None)  
+        build_pricing_table(json_data, table_data)  
+  
+    print(tabulate(table_data, headers='firstrow', tablefmt='psql'))  
+  
+if __name__ == "__main__":  
+    main()
