+++ bin/checkspotprice.sh
