+++ bin/create_azure_resources.sh
