+++ bin/enable_storage_key.sh
