+++ install.sh
@@ -1,43 +1,40 @@
-#!/bin/bash
-
-# Function to check if a command exists
-command_exists() {
-    command -v "$1" >/dev/null 2>&1
-}
-
-# Function to install a package if it's not already installed
-install_package() {
-    if ! command_exists "$1"; then
-        echo "Installing $1..."
-        sudo apt-get update && sudo apt-get install -y "$1"
-    else
-        echo "$1 is already installed."
-    fi
-}
-
-# Install Stow
-install_package stow
-
-# Clone dotfiles repository if it doesn't exist
-DOTFILES_DIR="$HOME/.dotfiles"
-if [ ! -d "$DOTFILES_DIR" ]; then
-    echo "Cloning dotfiles repository..."
-    git clone https://github.com/setuc/dotfiles.git "$DOTFILES_DIR"
-fi
-
-# Navigate to dotfiles directory
-cd "$DOTFILES_DIR" || exit
-
-# Use Stow to symlink configurations
-echo "Stowing configurations..."
-stow -v -R -t ~ bash
-stow -v -R -t ~ vim
-
-# Update .bashrc if necessary
-if ! grep -q "source ~/.bash_aliases" "$HOME/.bashrc"; then
-    echo 'if [ -f ~/.bash_aliases ]; then source ~/.bash_aliases; fi' >> "$HOME/.bashrc"
-fi
-
-# Additional setup steps can be added here
-
-echo "Dotfiles setup complete!"
+#!/bin/bash  
+  
+# Function to check if a command exists  
+command_exists() {  
+    command -v "$1" >/dev/null 2>&1  
+}  
+  
+# Function to install a package if it's not already installed  
+install_package() {  
+    if ! command_exists "$1"; then  
+        echo "Installing $1..."  
+        sudo apt-get update && sudo apt-get install -y "$1"  
+    else  
+        echo "$1 is already installed."  
+    fi  
+}  
+  
+# Install Stow  
+install_package stow  
+  
+# Clone dotfiles repository if it doesn't exist  
+DOTFILES_DIR="$HOME/.dotfiles"  
+if [ ! -d "$DOTFILES_DIR" ]; then  
+    echo "Cloning dotfiles repository..."  
+    git clone https://github.com/setuc/dotfiles.git "$DOTFILES_DIR"  
+fi  
+  
+# Navigate to dotfiles directory  
+cd "$DOTFILES_DIR" || exit  
+  
+# Use Stow to symlink configurations  
+echo "Stowing configurations..."  
+stow -v -R -t ~ bash  
+stow -v -R -t ~ bin  
+stow -v -R -t ~ oh-my-posh  
+  
+# Source the new .bashrc  
+source ~/.bashrc  
+  
+echo "Dotfiles setup complete!"
